moving storage functions to separate file
